feat: move error page to technical module
