+++ youtube-client/src/app/app.module.ts
@@ -5,11 +5,9 @@ import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { MaterialModule } from './material/material.module';
 import { CoreModule } from './core/core.module';
-import { ErrorPageComponent } from './error-page/error-page.component';
-import { ErrorComponent } from './components/svgs/error/error.component';
 
 @NgModule({
-  declarations: [AppComponent, ErrorPageComponent, ErrorComponent],
+  declarations: [AppComponent],
   imports: [AppRoutingModule, BrowserModule, MaterialModule, CoreModule, HttpClientModule],
   bootstrap: [AppComponent],
 })
