+++ youtube-client/src/app/technical/pages/error-page/error-page.component.scss
